+++ R/nn.R
@@ -65,7 +65,8 @@ nn_Module <- R6::R6Class(
           with_no_grad({
             param_applied <- fn(param)
           })
-          private$parameters_[[param_name]] <- nn_parameter(param_applied)
+          private$parameters_[[param_name]] <- 
+            nn_parameter(param_applied, param_applied$requires_grad)
         }
         
         if (!is.null(param) && !is_undefined_tensor(param$grad)) {
@@ -636,6 +637,11 @@ nn_module_list <- nn_module(
     x$.__enclos_env__$private$modules_[[y]]
   else
     NextMethod("[[")
+}
+
+#' @export
+as.list.nn_module_list <- function(x, ...) {
+  x$.__enclos_env__$private$modules_
 }
 
 #' @export