+++ tests/testthat/test-nn-rnn.R
@@ -234,10 +234,39 @@ test_that("rnn gpu", {
   rnn$to(device = "cuda")
   
   input <- torch_ones(1, 1, 10, device = "cuda")
-  out <- rnn(input)
+  
+  expect_message(out <- rnn(input), regexp = NA)
   
   expect_length(out, 2)
   expect_tensor_shape(out[[1]], c(1,1,1))
   expect_tensor_shape(out[[2]], c(1,1,1))
+  
+})
+
+test_that("lstm and gru works with packed sequences", {
+  # regression test for https://github.com/mlverse/torch/issues/499
+  
+  x <- torch_tensor(rbind(
+    c(1, 2, 0, 0),
+    c(1, 2, 3, 0),
+    c(1, 2, 3, 4)
+  ), dtype = torch_float())
+  x <- x[,,newaxis]
+  lens <- torch_tensor(c(2,3,4), dtype = torch_long())
+  
+  p <- nn_utils_rnn_pack_padded_sequence(x, lens, batch_first = TRUE, 
+                                         enforce_sorted = FALSE)
+  
+  rnn <- nn_lstm(1, 4)
+  out <- rnn(p)
+  
+  unpack <- nn_utils_rnn_pad_packed_sequence(out[[1]])
+  expect_tensor_shape(unpack[[1]], c(4, 3, 4))
+  
+  rnn <- nn_gru(1, 4)
+  out <- rnn(p)
+  
+  unpack <- nn_utils_rnn_pad_packed_sequence(out[[1]])
+  expect_tensor_shape(unpack[[1]], c(4, 3, 4))
   
 })